- Submitting a notify.module that will mail out a notification mail every X   minutes (configurable in the admin settings and filters page), to users that   have it enabled (requires the role to have the access).
Requires: the SQL to be added, and crontab.

Warning: this is still a work in progress. It works but does not have the
number of features I want it to have. Also the INSTALL file is not ready yet.
Will add and test more in the future once I get some more time. Feel free to
commit updates and patches.